// explore/explore-ui/src/app/modules/optimization-settings/constraints-settings/enums/constraint-option-type-key.enum.ts
 export enum ConstraintOptionTypeKey {
    NUMBER = 'constraintNumberValue',
    UNIT = 'ConstraintUnit',
    BOOLEAN = 'constraintBooleanValue',
    BOUNDS = 'ConstraintUpperLowerBounds',
    SECURITY_LIST = 'securityList',
    SECTOR = 'sectorConstraintType',
    BREAKDOWN = 'breakdownTree',
    FILTER = 'filter',
    BOUNDS_LONG_SHORT = 'ConditionalLongShortConstraintOption',
    QUICK_FACTOR_BLOCK = 'quickFactorBlock',
    FACTOR_TAG = 'factorTagList',
    CUSTOM_CALCULATION = 'customCalculation',
    CUSTOM_AGGREGATION = 'customAggregation',
    EFFICIENT_FRONTIER = 'efficientFrontierNumberValue',
    RELATIVE_ABSOLUTE = 'RelativeAbsolute',
    RELATIVE_PORT_BENCH = 'PortBench',
    RELATIVE_BOUNDS = 'RelativeUpperLowerBounds',
    EFFICIENT_ENABLED_RELATIVE_BOUNDS = 'EfficientEnabledRelativeBounds',
    EFFICIENT_ENABLED_BOUNDS = 'EfficientEnabledBounds',
    CUSTOM_TITLE = 'customColumnTitle',
    TEMP_ALIGNMENT_CLIMATE_SCENARIO = 'taClimateScenarioSettings',
    TRANSITION_CLIMATE_SCENARIO = 'tClimateScenarioSettings',
    ALLOW_SHORT_POSITIONS = 'allowShortPositionsOption',
    COLLAPSED_LOOK_THROUGH = 'collapsedLookthroughColumnOption',
    SELECTED_SECURITIES = 'selectedSecurities',
    SECURITY_CONSTRAINT_FILTER = 'securityConstraintFilter',
    SECURITY_CONSTRAINT_SUBSECTION = 'securityConstraintSubsection',
    MIN_TRADE_SIZE = 'minTradeSizeOptions',
    MIXED_INTEGER_SUPPORT = 'isMixedIntegerEnabled',
    STRESS_PNL_SCENARIO = 'stressPnlSetting',
    MISSING_DATA_HANDLING = 'ConstraintMissingDataHandling'
}